// src/utils/interfaces/episodeInterface.ts
export interface Episode {
    id: number;
    name: string;
    air_date: string;
    episode: string;
    characters: string[];
    url: string;
    created: string;
}

export interface EpisodesResponse {
    info: {
        count: number;
        pages: number;
        next: string;
        prev: string;
    };
    results: Episode[];
}